Massive change with how structs are being passed to JavaScript engine from C through WASM
Initially, I was defining an API interface for each different type to be printed out since C is a statically typed language. However, through some jugaad, I can now pass a pointer and iterate over a struct or array of structs by using the byte offsets for each struct.

This approach is a little brittle because changes in the struct in C require changes in the JS engine, but the API surface area is a lot smaller now.

Currently, the stack, the constants and the globals table are all printed out correctly for simple programs.